Rename RootComponent to RootPage for consistency
